refactor: extract insight fund flow category lists
